Extract reload policy handler to a new file.
